perf: refactor authenticated enterprise app routes to eliminate duplicate work
By moving the authenticated enterprise app routes further down the component tree (conversely, moving the `AuthenticatedUserSubsidyPage` component higher in the component tree), we are ensuring the common business logic between pages is not duplicated between page route changes, for example when the user goes from the Course Dashboard page to the Search page.
