+++ src/components/program-progress/index.js
@@ -1,1 +1,2 @@
 export { default as ProgramListingPage } from './ProgramListingPage';
+export { default as ProgramProgressPage } from './ProgramProgressPage';
